remove redundant hardcoded taxonomies from models; rename type to name to avoid shadowing
This commit removes the hardcoded `choices` from the Django models to avoid unnecessary redundancy. The tables will be filled by either fixtures or a migration file, and since these are taxonomies they will generally be static. As part of this, I also renamed the `type` column in OrgType and SchoolType to avoid shadowing the `type` keyword in Python.